flask_api/requirements.txt
flask_mysqldb
aspose.pdf
flask
flask_cors